Improved util behaviour. Conversion: No longer deletes files that were already .mp3. Prefixing: Changed separator because it caused songs to be played out of order sometimes.

+++ util/convert_files.sh
@@ -1,6 +1,9 @@
 #!/bin/bash
 
 for filename in *; do
-	ffmpeg -i "$filename" -vn -ar 44100  -ac 2 -b:a 320k "${filename%.*}.mp3"
-	rm "$filename"
+	new="${filename%.*}.mp3"
+	ffmpeg -i "$filename" -vn -ar 44100  -ac 2 -b:a 320k "$new"
+	if [ "$new" != "$filename" ]; then
+		rm "$filename"
+	fi
 done
